Seperate post closure babel pass from compile
